+++ src/meta.py
@@ -161,7 +161,7 @@ class Parameter(Object):
     scale_factor = Float.T(default=1., optional=True)
     scale_unit = Unicode.T(optional=True)
     label = Unicode.T(optional=True)
-    optional = Bool.T(default=False, optional=True)
+    optional = Bool.T(default=True, optional=True)
 
     def __init__(self, *args, **kwargs):
         if len(args) >= 1: